Added check of normalized affiliation attribution for supplementary publications extracted from HAL database
