+++ bmfuncts/__init__.py
@@ -16,6 +16,7 @@ from bmfuncts.add_ifs import *
 from bmfuncts.use_homonyms import *
 from bmfuncts.use_otps import *
 from bmfuncts.update_employees import *
+from bmfuncts.build_pub_authors import *
 from bmfuncts.merge_pub_employees import *
 from bmfuncts.consolidate_pub_list import *
 from bmfuncts.update_impact_factors import *
